Moved actor.py into own package dir.
